Updated weather data retrieval scripts
